discord_bot/requirements.txt
discord.py>=2.3.0
stripe>=7.0.0
python-dotenv>=1.0.0
